+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@lindeneg/server",
+    "name": "lindeneg.org",
     "version": "1.0.0",
     "type": "module",
     "private": true,
